+++ bzt/__init__.py
@@ -46,7 +46,13 @@ class TaurusInternalException(TaurusException):
 
 
 class ToolError(TaurusException):
-    pass
+    def __init__(self, message, diagnostics=None):
+        """
+        :type message: str
+        :type diagnostics: list[str]
+        """
+        super(ToolError, self).__init__(message)
+        self.diagnostics = diagnostics
 
 
 class TaurusNetworkError(TaurusException):