+++ tests/modules/test_Siege.py
@@ -115,6 +115,21 @@ class TestSiegeExecutor(BZTestCase):
         obj.post_process()
         self.assertNotEquals(obj.process, None)
 
+    def test_diagnostics(self):
+        obj = SiegeExecutor()
+        obj.engine = EngineEmul()
+        obj.settings.merge({"path": get_res_path(TOOL_NAME),})
+        obj.execution.merge({"iterations": 1, "scenario": {"requests": ["http://blazedemo.com"]}})
+        obj.prepare()
+        try:
+            obj.startup()
+            while not obj.check():
+                time.sleep(obj.engine.check_interval)
+        finally:
+            obj.shutdown()
+        obj.post_process()
+        self.assertIsNotNone(obj.get_error_diagnostics())
+
 
 class TestDataLogReader(BZTestCase):
     def test_read(self):